<IMP> Clock modifications to get to 4 kHz processing

+++ lib/Filter/src/Filter.c
@@ -5,19 +5,19 @@
 #include <math.h>
 
 
-void Filter_init(Filter* f, const float num[], const float den[], const float x0[], const float y0[], 
+void Filter_init(Filter* f, const double num[], const double den[], const double x0[], const double y0[], 
                    const unsigned num_x_sz, const unsigned den_y_sz) 
 {
-    memcpy(f->num, num, num_x_sz*sizeof(float));
-    memcpy(f->den, den, den_y_sz*sizeof(float));
-    memcpy(f->x, x0, num_x_sz*sizeof(float));
-    memcpy(f->y, y0, den_y_sz*sizeof(float));
+    memcpy(f->num, num, num_x_sz*sizeof(double));
+    memcpy(f->den, den, den_y_sz*sizeof(double));
+    memcpy(f->x, x0, num_x_sz*sizeof(double));
+    memcpy(f->y, y0, den_y_sz*sizeof(double));
     f->num_x_sz = num_x_sz;
     f->den_y_sz = den_y_sz;
 }
 
-float Filter_input(Filter *f, float point) {
-    float *num = f->num, *den = f->den, *x = f->x, *y = f->y;
+double Filter_input(Filter *f, double point) {
+    double *num = f->num, *den = f->den, *x = f->x, *y = f->y;
     unsigned num_x_sz = f->num_x_sz, den_y_sz = f->den_y_sz;
 
     for (int i = 1; i<num_x_sz; ++i) {  // pop front and ...
@@ -37,26 +37,26 @@ float Filter_input(Filter *f, float point) {
     // https://en.wikipedia.org/wiki/Kahan_summation_algorithm#Further_enhancements
     /*
     // This was not used because the processing cost was too much for keeping Fs >= 2kHz.
-    float sum = 0;
-    float c = 0;
+    double sum = 0;
+    double c = 0;
     for (int i = 0; i<num_x_sz; ++i) {  // Difference equation sum for x
-        float sum_aux = fmal(num[i]/den[0], x[num_x_sz-1 - i], sum);
+        double sum_aux = fma(num[i]/den[0], x[num_x_sz-1 - i], sum);
 
-        if (fabsl(sum) > fabsl(num[i]/den[0] * x[num_x_sz-1 - i]))
-            c += fmal(num[i]/den[0], x[num_x_sz-1 - i], sum-sum_aux);
+        if (fabs(sum) > fabs(num[i]/den[0] * x[num_x_sz-1 - i]))
+            c += fma(num[i]/den[0], x[num_x_sz-1 - i], sum-sum_aux);
         else
-            c += fmal(num[i]/den[0], x[num_x_sz-1 - i], -sum_aux) + sum;
+            c += fma(num[i]/den[0], x[num_x_sz-1 - i], -sum_aux) + sum;
 
         sum = sum_aux;
     }
 
     for (int i = 1; i<den_y_sz; ++i) {  // Difference equation sum for y
-        float sum_aux = fmal(-den[i]/den[0], y[den_y_sz-1 - i], sum);
+        double sum_aux = fma(-den[i]/den[0], y[den_y_sz-1 - i], sum);
 
-        if (fabsl(sum) > fabsl(den[i]/den[0] * y[den_y_sz-1 - i]))
-            c += fmal(-den[i]/den[0], y[den_y_sz-1 - i], sum-sum_aux);
+        if (fabs(sum) > fabs(den[i]/den[0] * y[den_y_sz-1 - i]))
+            c += fma(-den[i]/den[0], y[den_y_sz-1 - i], sum-sum_aux);
         else
-            c += fmal(-den[i]/den[0], y[den_y_sz-1 - i], -sum_aux) + sum;
+            c += fma(-den[i]/den[0], y[den_y_sz-1 - i], -sum_aux) + sum;
 
         sum = sum_aux;
     }
@@ -64,11 +64,11 @@ float Filter_input(Filter *f, float point) {
     */
 
     for (int i = 0; i<num_x_sz; ++i) {  // Difference equation sum for x
-        y[den_y_sz-1] = fmaf(num[i]/den[0], x[num_x_sz-1 - i], y[den_y_sz-1]);
+        y[den_y_sz-1] = fma(num[i]/den[0], x[num_x_sz-1 - i], y[den_y_sz-1]);
     }
 
     for (int i = 1; i<den_y_sz; ++i) {  // Difference equation sum for y
-        y[den_y_sz-1] = fmaf(-den[i]/den[0], y[den_y_sz-1 - i], y[den_y_sz-1]);
+        y[den_y_sz-1] = fma(-den[i]/den[0], y[den_y_sz-1 - i], y[den_y_sz-1]);
     }
 
     return y[den_y_sz-1];   // return the new output point (y_n)
