+++ lib/Filter/src/Filter.h
@@ -2,15 +2,15 @@
 #define FILTER_HEADER_GUARD
 
 typedef struct Filter {
-    float x[16], y[16];
-    float num[16], den[16];
+    double x[16], y[16];
+    double num[16], den[16];
     unsigned num_x_sz, den_y_sz;
 } Filter;
 
 
-void Filter_init(Filter* f, const float num[], const float den[], const float x0[], const float y0[],
+void Filter_init(Filter* f, const double num[], const double den[], const double x0[], const double y0[],
                    const unsigned num_x_sz, const unsigned den_y_sz);
 
-float Filter_input(Filter *f, float point);
+double Filter_input(Filter *f, double point);
 
 #endif
